Added the new filter types
From now on this library will only be focusing on sensor fusion for 6 DoF
and 9 DoF systems. These new filters will only use Madgwick's math to
operate. This will help to make this library simpler to work with. Note that
as of this commit, the filter functions still need to be implemented.